Add support for error messages to fieldsets.
This adds support for error messages for radio buttons and check
boxes, plus other tweaks to the HTML rendering.

Also refactors how HTML is generated for these fieldsets and how the
specs tests the related methods. The goal is to move away from brittle
tests and to treat them more as a spec and less like a full regression
test.