Remove duplicated name input field html expectation
Add a method to generate expected html output for a name
input field.

Replace duplicated name input field html expectations in the
form builder spec with calls to the method. Change parameters
to change the expected html in the small number of places it
varies.

This approach keep the html expectation as html text. The intention
is to make it easy for developers doing frontend to see the
generated markup in the test expectation.

Some duplication in test expectations is acceptable when it leads
to expectations that are concise and easy to read and reason about.

An alternate pull request to remove this duplication created a
FormGroupObject model to perform tests. This different approach
lost having the expected html output easily readable in a few lines
of code.